Worked around a bug in iOS 7.1 where the min and max tracks of UISliders jump around if you set the tint color. I think I ended up with something nicer by setting the bg color instead of the tint color. Fixed the warnings that Xcode 5.1 gave me. Fixed the bug where I was getting 'no index path for table cell being reused' on the header when there are no records in the database.

+++ MoodLog/MlChartCollectionViewController.m
@@ -34,7 +34,7 @@ Boolean firstLoad;
     for (index = 0, numberOfLines = 0; index < stringLength; numberOfLines++)
         index = NSMaxRange([self.text lineRangeForRange:NSMakeRange(index, 0)]);
 
-    NSUInteger newLinesToAdd = abs(labelLines - numberOfLines - 1); // abs is just a stab in the dark
+    NSUInteger newLinesToAdd = (int)abs((int)labelLines - (int)numberOfLines - 1); // abs is just a stab in the dark
     for(int i=0; i<newLinesToAdd; i++)
         self.text = [NSString stringWithFormat:@" \n%@",self.text];
 }
